\name{btable}
\alias{btable}
\docType{data}
\title{b-Table File for Shell Data}
\description{
\file{btable.txt} is an alias for \file{dsi203_bmax4000}, a (nx4) b-table data file used in conjunction with the data set \file{data.nii.gz} (alias "DSI 203-point 3mm"), as provided by the \emph{Advanced Biomedical MRI Lab, National Taiwan University Hospital}, and referenced in the package \pkg{DSI_Studio} developed by Fang-Cheng Yeh.
The first column specifies the b-values, and the remaining 3 columns specify the b-vectors.

\file{btable.txt}
The sampling points in \file{btable.txt} (\file{dsi203_bmax4000}) are distributed in a S2 shell with 203 points. The b-values range from 0 to 4000. 
}

\references{
Yeh, F.-C., Wedeen, V. J., and Tseng, W.-Y. I. Generalized q-
Sampling Imaging. \emph{IEEE Transactions on Medical Imaging} 29, 9 (2010),
1626--1635.

NITRC repository,  \url{http://www.nitrc.org}.
}

\keyword{ data }
