\name{dsi203_bmax4000.Rd}
\alias{dsi203_bmax4000.Rd}
\docType{data}
\title{DSI b-Table File}
\description{
3D grid b-table file used in diffusion data acquisition for data set \file{data.nii.gz} (alias "DSI 203-point 3mm"), as provided by the \emph{Advanced Biomedical MRI Lab, National Taiwan University Hospital}, and referenced in the package \pkg{DSI_Studio} developed by Fang-Cheng Yeh.
}

\references{
Yeh, F.-C., Wedeen, V. J., and Tseng, W.-Y. I. Generalized q-
Sampling Imaging. \emph{IEEE Transactions on Medical Imaging} 29, 9 (2010),
1626--1635.

NITRC repository,  \url{http://www.nitrc.org}.
}

\keyword{ data }
