\name{gdimap-package}
\Rdversion{0.0-3}
\alias{gdimap-package}
\docType{package}
\title{
Generalized Diffusion Magnetic Resonance Imaging
}

\description{
The package implements algorithms to estimate and visualize the orientation of neuronal pathways in model-free methods (q-space imaging methods).
For fibre orientation estimation based on mixtures of von Mises-Fischer (vMF) distributions see \code{\link{gqi.odfvmflines}}, and \code{\link{gqi.odfvmf}}.
For fibre orientation estimation based on local peak detection procedures see \code{\link{gqi.odfpeaklines}}, and \code{\link{gqi.odfpeaks}}.
Visualization of RGB-maps and glyph-maps is implemented through \code{\link{rgbvolmap}} and \code{\link{gqi.odfvxgrid}}, respectively.
Various types of simulations of diffusion and ODF profiles, illustrating the application of vMF mixtures for fibre orientation estimation are implemented in \code{\link{simulglyph.vmf}}, \code{\link{simul.simplefield}}, \code{\link{simul.fandtasia}}.
Data sets used in the examples are documented.
}

\details{
For a complete list of functions type \code{library(help=gdimap)}.
}

\author{ Adelino Ferreira da Silva, Universidade Nova de Lisboa,
  Faculdade de Ciencias e Tecnologia, Portugal, \email{afs at fct.unl.pt}
}

\keyword{ package }

